Moved IdlingIncident and Location into its own models file
