+++ app/models/__init__.py
@@ -4,3 +4,4 @@ module (as opposed to just their python files)
 """
 
 from user import *  # noqa
+from idling_incident import *  # noqa
